requirements.txt
ably==2.0.6
amqp==5.2.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.1
daphne==4.1.2
defusedxml==0.8.0rc2
distlib==0.3.8
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==4.2.16
django-allauth==65.0.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cleanup==9.0.0
django-configurations==2.5.1
django-cors-headers==4.4.0
django-filter==24.3
django-htmx==1.19.0
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==29.0.0
filelock==3.15.4
google-auth==2.35.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
methoddispatch==3.0.2
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyee==11.1.1
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-django==4.9.0
pytest-factoryboy==2.7.0
pytest-mock==3.14.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
rsa==4.9
service-identity==24.1.0
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
spotipy==2.16.1
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
virtualenv==20.26.6
virtualenv2==0.0.1
wcwidth==0.2.13
websockets==12.0
youtube-search==1.1.1
zope.interface==7.1.0
